+++ lib/src/logger.dart
@@ -2,14 +2,17 @@ import 'package:ansicolor/ansicolor.dart';
 
 enum Severity { debug, success, info, warning, error }
 
+/// Logger class that allows simple logging of messages, including the origin of the message.
 class Logger {
+  /// Name of the origin of messages, for example the `Network` class will have messaged displayed
+  /// with a `Network` prefix.
   final String name;
 
   const Logger(this.name);
 
-  // Assigns a colour to a severity and outputs a message once formatted
+  /// Assigns a colour to the severity of the message and outputs a coloured message once formatted
   Future log(dynamic message, {Severity severity = Severity.info}) async {
-    AnsiPen pen;
+    final AnsiPen pen;
 
     switch (severity) {
       case Severity.debug:
@@ -25,9 +28,7 @@ class Logger {
         pen = AnsiPen()..yellow();
         break;
       case Severity.error:
-        pen = AnsiPen()
-          ..red()
-          ..yellow();
+        pen = AnsiPen()..red()..yellow();
         break;
     }
 
@@ -36,12 +37,14 @@ class Logger {
     print('[$name] ~ $message');
   }
 
-  // Interfaces for the log command
+  /// Prints a debug message
   Future debug(dynamic message) async => log(message, severity: Severity.debug);
-  Future success(dynamic message) async =>
-      log(message, severity: Severity.success);
+  /// Prints a success message
+  Future success(dynamic message) async => log(message, severity: Severity.success);
+  /// Prints an informational message
   Future info(dynamic message) async => log(message, severity: Severity.info);
-  Future warning(dynamic message) async =>
-      log(message, severity: Severity.warning);
+  /// Prints a warning message
+  Future warning(dynamic message) async => log(message, severity: Severity.warning);
+  /// Prints an error message
   Future error(dynamic message) async => log(message, severity: Severity.error);
 }