+++ lib/synadart.dart
@@ -2,4 +2,4 @@ library synadart;
 
 export 'src/activation.dart';
 
-export 'src/networks/multilayer_perceptron.dart';
+export 'src/networks/deep_feedforward.dart';